requirements.txt
google-cloud-texttospeech
python-telegram-bot
beautifulsoup4
requests
pandas